+++ R/SIG.conversion.l93dec2.R
@@ -1,6 +1,6 @@
 #' Conversion de coordonnées en Lambert 93 vers degrés décimaux
 #'
-#' Cette fonction permet de convertir des coordonnées en du Lambert 93 vers degrés décimaux
+#' Cette fonction permet de convertir des coordonnées en du Lambert 93 (2154) vers degrés décimaux (4326)
 #' @name SIG.conversion.l93dec2
 #' @param coord_x Coordonnée x à convertir
 #' @param coord_y Coordonnée y à convertir
@@ -11,18 +11,15 @@
 #' @examples
 #' SIG.conversion.l93dec2(888227, 6655627)
 
-    SIG.conversion.l93dec2 <- function(coord_x, coord_y) {
-      
-      wgs84 <-
-        tribble(~id, ~point_coord_x, ~point_coord_y,
-              1, coord_x, coord_y
-      ) %>% 
-        st_as_sf(coords = c("point_coord_x", "point_coord_y"), crs = 2154, dim = "XY", remove = F) %>%
-        st_transform(4326) %>% 
-        mutate(chsta_coord_x_conv = st_coordinates(.)[,1]) %>% 
-        mutate(chsta_coord_y_conv= st_coordinates(.)[,2]) %>% 
-        st_drop_geometry() %>% 
-        st_as_sf(coords = c("chsta_coord_y_conv", "chsta_coord_x_conv"), crs = 4326, dim = "XY", remove = T)
-      
-      return(wgs84)
-    }
+    SIG.conversion.l93dec2 <- function(coord_x = NA, coord_y = NA) {
+        
+        #### Test de cohérence ####
+        if(is.na(coord_x)) stop("Pas de coord_x fournie")
+        if(is.na(coord_y)) stop("Pas de coord_y fournie")
+        
+        #### Calcul ####
+        cible <- SIG.conversion(coord_x, coord_y, 2154, 4326)
+        
+        #### Sortie ####
+        return(cible)
+      }