+++ R/chronique.analyse.R
@@ -17,7 +17,7 @@
 
 chronique.analyse <- function(
   data = data,
-  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique", "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée", "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie", "Pluviométrie"),
+  typemesure = c("Thermie", "Thermie barométrique", "Thermie piézométrique", "Barométrie", "Piézométrie", "Piézométrie brute", "Piézométrie compensée", "Piézométrie calée", "Piézométrie NGF", "Oxygénation", "Hydrologie", "Pluviométrie", "Nitrates", "Conductivité"),
   pasdetemps = 1,
   seuils = c(25,22,19,15,4),
   seuilexcesdefaut = 12,